added activities adding + loading states

+++ src/components/contact-page-table-edit.tsx
@@ -16,6 +16,7 @@ import {
 } from "@/components/ui/tooltip";
 import { useState } from "react";
 import { api } from "~/utils/api";
+import { toast } from "sonner";
 
 export const ContactPageTableEdit: React.FC<{ contactId: string }> = ({
   contactId,
@@ -29,6 +30,7 @@ export const ContactPageTableEdit: React.FC<{ contactId: string }> = ({
       setDeleteLoading(true);
     },
     onSuccess: () => {
+      toast("Contact deleted succesfully.");
       void ctx.contact.getAll.invalidate();
     },
     onError: () => {
@@ -44,27 +46,27 @@ export const ContactPageTableEdit: React.FC<{ contactId: string }> = ({
       }}
     >
       <div key={`cpte-${contactId}`} className="flex">
-        <div className="box-content h-4 w-4 cursor-pointer rounded-l-md border p-2 text-red-500 transition-colors hover:bg-accent">
-          {!deleteLoading ? (
-            <TooltipProvider delayDuration={300}>
-              <Tooltip>
-                <TooltipTrigger asChild>
-                  <Trash
-                    className="h-4 w-4"
-                    onClick={() => {
-                      deleteContact({ id: contactId });
-                    }}
-                  />
-                </TooltipTrigger>
-                <TooltipContent className="mb-2">
-                  <p>Delete</p>
-                </TooltipContent>
-              </Tooltip>
-            </TooltipProvider>
-          ) : (
-            <Loader2 className="h-4 w-4 animate-spin" />
-          )}
-        </div>
+        <TooltipProvider delayDuration={300}>
+          <Tooltip>
+            <TooltipTrigger asChild>
+              <div
+                className="box-content h-4 w-4 cursor-pointer rounded-l-md border p-2 text-red-500 transition-colors hover:bg-accent"
+                onClick={() => {
+                  !deleteLoading && deleteContact({ id: contactId });
+                }}
+              >
+                {!deleteLoading ? (
+                  <Trash className="h-4 w-4" />
+                ) : (
+                  <Loader2 className="h-4 w-4 animate-spin" />
+                )}
+              </div>
+            </TooltipTrigger>
+            <TooltipContent className="">
+              <p>Delete</p>
+            </TooltipContent>
+          </Tooltip>
+        </TooltipProvider>
         <div className="box-content cursor-pointer rounded-none border-b border-t p-2 text-muted-foreground transition-colors hover:bg-accent">
           <Link className="h-4 w-4" />
         </div>
